requirements.txt
gunicorn
django==1.6.5
djangorestframework
django-treebeard
django-oauth2-provider